[maven-release-plugin] copy for tag fablab-manager-2.1.0

+++ src/main/java/net/collaud/fablab/manager/dao/GroupRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Query;
  *
  * @author Gaetan Collaud <gaetancollaud@gmail.com>
  */
-public interface GroupRepository extends JpaRepository<GroupEO, Integer>{
+public interface GroupRepository extends JpaRepository<GroupEO, Long>{
 	
 	@Query("SELECT DISTINCT g FROM GroupEO g LEFT JOIN FETCH g.roles")
 	List<GroupEO> findAllWithRoles();
